Move code generation into subcrate.
