+++ Cargo.toml
@@ -12,7 +12,7 @@ exclude = [".gitignore", ".travis.yml", "docgen.sh"]
 build = "build.rs"
 
 [features]
-default = ["serde_codegen"]
+default = ["rusoto_codegen/serde_codegen"]
 nightly = ["serde_macros"]
 all = ["dynamodb", "ecs", "kms", "s3", "sqs"]
 dynamodb = []
@@ -21,19 +21,12 @@ kms = []
 s3 = []
 sqs = []
 
-[build-dependencies]
-regex = "^0.1.51"
-Inflector = "0.2.0"
-# for faster development of botocore_parser this can can be set to use a local copy:
-botocore_parser = { git = "https://github.com/rusoto/botocore_parser.git" }
-serde_codegen = { version = "0.6.14", optional = true }
-serde_json = "0.6.0"
-syntex = "^0.29.0"
+[build-dependencies.rusoto_codegen]
+path = "codegen"
 
 [dependencies]
 serde = "0.6.15"
 serde_json = "0.6.0"
-serde_macros = { version = "0.6.14", optional = true }
 xml-rs = "^0.1.26"
 time = "^0.1.34"
 openssl = "^0.7.5"
@@ -44,6 +37,7 @@ regex = "^0.1.51"
 chrono = "^0.2.19"
 rust-crypto = "^0.2.34"
 log = "^0.3.3"
+serde_macros = { version = "0.7.0", optional = true }
 
 [dev-dependencies]
 env_logger = "^0.3.2"
