+++ src/lib.rs
@@ -1,5 +1,7 @@
 #![crate_name = "rusoto"]
 #![crate_type = "lib"]
+#![cfg_attr(feature = "nightly", feature(custom_derive, plugin))]
+#![cfg_attr(feature = "nightly", plugin(serde_macros))]
 #![allow(dead_code)]
 
 //! # Rusoto
